requirements.txt
apscheduler==3.10.4
feedparser==6.0.11
trafilatura==1.12.2
requests==2.32.3
weasyprint==62.3
jinja2==3.1.4
python-dateutil==2.9.0
imap-tools==1.7.2
pyyaml==6.0.2
openai==1.54.3          # for AI summaries (OpenAI-compatible)
